# backend/openapi/openapi-v0.1.0.yaml
openapi: 3.0.3
info:
  title: Task Tracker API
  version: "0.1.0"
  description: |
    Контракты Backend/Frontend для таск-трекера.
    Документ ориентирован на использование в Swagger UI / Redoc и генерацию SDK.

    Базовый префикс всех эндпоинтов: /api/v1
servers:
  - url: http://localhost:8000/api/v1
    description: Основной сервер API
...
security:
  - bearerAuth: []
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [Projects]
      summary: Список проектов текущего пользователя
      description: |
        Возвращает проекты, где пользователь является участником (или создателем).
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Список проектов с пагинацией
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedProjects"
    post:
      operationId: createProject
      tags: [Projects]
      summary: Создать проект
      description: |
        Создаёт проект с уникальным key (например PAY).
        Обычно доступно только пользователям с соответствующими правами (в MVP — любому авторизованному).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Проект создан
          content:
            application/json:
              schema:
                type: object
                required: [project]
                properties:
                  project:
                    $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [Projects]
      summary: Получить проект
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Проект
          content:
            application/json:
              schema:
                type: object
                required: [project]
                properties:
                  project:
                    $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{projectId}/onboarding/recommendations:
    get:
      operationId: getOnboardingRecommendations
      tags: [Projects]
      summary: Получить onboarding-рекомендации по проекту
      description: |
        Возвращает рекомендации для onboarding на основе проектных данных:
        - reads (Wiki pages)
        - issues_to_review (Issues)
        - key_people (Users)
        Для каждого элемента возвращается reason и опциональный score.
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: force
          in: query
          required: false
          description: Принудительно пересчитать рекомендации, игнорируя кэш.
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Onboarding-рекомендации
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingRecommendationsResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{projectId}/me/preferences:
    get:
      operationId: getMyProjectPreferences
      tags: [Projects]
      summary: Получить персональные настройки проекта
      description: |
        Возвращает настройки текущего пользователя для конкретного проекта.
        Сейчас используется для флага new_employee_mode.
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Настройки текущего пользователя
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectOnboardingPreferencesResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateMyProjectPreferences
      tags: [Projects]
      summary: Обновить персональные настройки проекта
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectOnboardingPreferencesRequest"
      responses:
        "200":
          description: Настройки обновлены
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectOnboardingPreferencesResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{projectId}/onboarding/assignees:
    get:
      operationId: getProjectOnboardingAssignees
      tags: [Projects]
      summary: Получить список назначенных onboarding-пользователей
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Пользователи, которым назначен onboarding-режим
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectOnboardingAssigneesResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateProjectOnboardingAssignees
      tags: [Projects]
      summary: Назначить onboarding-пользователей проекта
      description: |
        Обновляет полный список участников, которым показывается New Employee onboarding.
        Требует роль admin_project.
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectOnboardingAssigneesRequest"
      responses:
        "200":
          description: Список назначений обновлен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectOnboardingAssigneesResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{projectId}/users:
    get:
      operationId: listProjectUsers
      tags: [Projects]
      summary: Список участников проекта
      description: |
        Возвращает участников и их роли. Используется в экране "Участники проекта".
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Участники проекта
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedProjectUsers"
    post:
      operationId: addProjectUser
      tags: [Projects]
      summary: Добавить пользователя в проект
      description: |
        Требует роль project_admin в данном проекте.
        В MVP инвайт выполняется по user_id (без email-приглашений).
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddProjectUserRequest"
      responses:
        "201":
          description: Пользователь добавлен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectUser"
        "403":
          $ref: "#/components/responses/Forbidden"
  /projects/{projectId}/users/{userId}:
    delete:
      operationId: removeProjectUser
      tags: [Projects]
      summary: Удалить пользователя из проекта
      description: |
        Требует роль project_admin.
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/UserId"
      responses:
        "204":
          description: Пользователь удалён
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT access token в заголовке Authorization: Bearer <token>"
  parameters:
    Limit:
      name: limit
      in: query
      description: Максимум элементов на странице (пагинация).
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Offset:
      name: offset
      in: query
      description: Смещение от начала списка (пагинация).
      schema:
        type: integer
        minimum: 0
        default: 0
...
    UserId:
      name: userId
      in: path
      required: true
      description: UUID пользователя.
      schema:
        $ref: "#/components/schemas/UUID"
    ProjectId:
      name: projectId
      in: path
      required: true
      description: UUID проекта.
      schema:
        $ref: "#/components/schemas/UUID"
...
  responses:
    BadRequest:
      description: Ошибка запроса / ошибка валидации
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Forbidden:
      description: Нет прав
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Не найдено
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Конфликт
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    UUID:
      type: string
      format: uuid
      description: UUID v4.
      example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
...
    DateTime:
      type: string
      format: date-time
      description: ISO 8601 datetime (с таймзоной).
      example: "2026-03-02T10:15:30Z"
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"
...
    Project:
      type: object
      required: [id, key, name, created_by, created_at]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        key:
          type: string
          maxLength: 20
          description: Ключ проекта (для ключей задач).
          example: PAY
        name:
          type: string
          description: Название.
        created_by:
          $ref: "#/components/schemas/UUID"
          description: Создатель.
        created_at:
          $ref: "#/components/schemas/DateTime"
    CreateProjectRequest:
      type: object
      required: [key, name]
      properties:
        key:
          type: string
          minLength: 2
          maxLength: 20
          description: Уникальный ключ проекта.
        name:
          type: string
          minLength: 1
          maxLength: 200
          description: Название проекта.
...
    ProjectUser:
      type: object
      required: [project_id, user_id, role]
      properties:
        project_id:
          $ref: "#/components/schemas/UUID"
        user_id:
          $ref: "#/components/schemas/UUID"
        role:
          $ref: "#/components/schemas/ProjectRole"
...
    AddProjectUserRequest:
      type: object
      required: [user_id, role]
      properties:
        user_id:
          $ref: "#/components/schemas/UUID"
        role:
          $ref: "#/components/schemas/ProjectRole"
...
    PagedProjects:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        total:
          type: integer
    PagedProjectUsers:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ProjectUserExpanded"
        total:
          type: integer
...
    OnboardingReadItem:
      type: object
      required: [page_id, title, reason]
      properties:
        page_id:
          $ref: "#/components/schemas/UUID"
        title:
          type: string
        reason:
          type: string
        score:
          type: number
          format: double
          nullable: true
...
    OnboardingRecommendationsResponse:
      type: object
      required: [reads, issues_to_review, key_people, generated_at, cached]
      properties:
        reads:
          type: array
          items:
            $ref: "#/components/schemas/OnboardingReadItem"
        issues_to_review:
          type: array
          items:
            $ref: "#/components/schemas/OnboardingIssueToReviewItem"
        key_people:
          type: array
          items:
            $ref: "#/components/schemas/OnboardingKeyPersonItem"
        generated_at:
          $ref: "#/components/schemas/DateTime"
        cached:
          type: boolean
          description: true, если ответ взят из кэша в пределах TTL.
...
    UpdateProjectOnboardingPreferencesRequest:
      type: object
      required: [new_employee_mode]
      properties:
        new_employee_mode:
          type: boolean
    ProjectOnboardingPreferencesResponse:
      type: object
      required: [preferences]
      properties:
        preferences:
          $ref: "#/components/schemas/ProjectOnboardingPreferences"
...
    ProjectOnboardingAssigneesResponse:
      type: object
      required: [assignees]
      properties:
        assignees:
          type: array
          items:
            $ref: "#/components/schemas/OnboardingAssigneeItem"
    UpdateProjectOnboardingAssigneesRequest:
      type: object
      required: [user_ids]
      properties:
        user_ids:
          type: array
          items:
            $ref: "#/components/schemas/UUID"